requirements.txt
google-genai
exa_py
streamlit
streamlit-mermaid
plotly
chromadb
pymupdf
pillow
python-dotenv
gTTS
markitdown